+++ CommandLine/emake-tests/Parsing/parser-test.cpp
@@ -1,6 +1,7 @@
 #include <languages/lang_CPP.h>
 #include <parsing/ast.h>
 #include <parsing/parser.cpp>
+#include "matchers.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
@@ -36,243 +37,6 @@ struct ParserTester {
 void assert_identifier_is(AST::Node *node, std::string_view name) {
   ASSERT_EQ(node->type, AST::NodeType::IDENTIFIER);
   ASSERT_EQ(node->As<AST::IdentifierAccess>()->name.content, name);
-}
-
-std::string ExpectedMsg = "";
-
-MATCHER_P(IsDeclaration, decls, "") {
-  bool b1 = arg->type == AST::NodeType::DECLARATION;
-  if (!b1) {
-    ExpectedMsg = "From IsDeclaration Matcher: NodeType = DECLARATION\n";
-    *result_listener << "got NodeType = " << AST::NodeToString(arg->type) << "\n";
-    return 0;
-  }
-
-  auto *decl = arg->template As<AST::DeclarationStatement>();
-  bool b2 = decl->storage_class == AST::DeclarationStatement::StorageClass::TEMPORARY,
-       b3 = decl->declarations.size() == decls.size();
-
-  if (!b2 || !b3) {
-    ExpectedMsg = "From IsDeclaration Matcher: ";
-
-    if (!b2) {
-      ExpectedMsg += "StorageClass = TEMPORARY\n";
-      *result_listener << "got StorageClass = " << AST::DeclarationStatement::StorageToString(decl->storage_class)
-                       << "\n";
-    }
-    if (!b3) {
-      ExpectedMsg += "DeclarationsSize = " + to_string(decls.size()) + "\n";
-      *result_listener << "got DeclarationsSize = " << to_string(decl->declarations.size()) << "\n";
-    }
-  }
-
-  bool b4 = 1;
-  for (size_t i = 0; i < decls.size(); i++) {
-    auto &decli = decl->declarations[i].declarator->decl;
-    b4 = b4 && decl->declarations[i].init != nullptr;
-    b4 = b4 && decl->declarations[i].declarator->def == jdi::builtin_type__int;  // need to send the type
-    b4 = b4 && decl->declarations[i].declarator->flags == 0;
-    b4 = b4 && decli.name.content == decls[i];
-    b4 = b4 && decli.components.size() == 0;
-    if (!b4) {
-      if (ExpectedMsg == "") ExpectedMsg = "From IsDeclaration Matcher: ";
-      ExpectedMsg += "Declaration [" + to_string(i) +
-                     "] has init != nullptr, def = jdi::builtin_type__int, flags = 0, name.content = " + decls[i] +
-                     ", components.size() = 0\n";
-      *result_listener << " got Declaration [" << to_string(i) << "] has init "
-                       << ((decl->declarations[i].init) ? "!=" : "=")
-                       << " nullptr, def = jdi::builtin_type__int, flags = "
-                       << to_string(decl->declarations[i].declarator->flags)
-                       << ", name.content = " << decli.name.content
-                       << ", components.size() = " << to_string(decli.components.size()) << "\n";
-    }
-  }
-
-  return b2 && b3 && b4;
-}
-
-MATCHER_P2(IsCast, cast_kind, expr_type, "") {
-  bool b1 = arg->type == AST::NodeType::CAST;
-  if (!b1) {
-    ExpectedMsg = "From IsCast Matcher: NodeType = CAST\n";
-    *result_listener << "got NodeType = " << AST::NodeToString(arg->type) << "\n";
-    return 0;
-  }
-
-  auto *cast = arg->template As<AST::CastExpression>();
-  auto *expr = cast->expr->template As<AST::Node>();
-
-  bool b2 = cast->ft.def == jdi::builtin_type__int, b3 = cast->ft.flags == 0, b4 = cast->ft.decl.components.size() == 0,
-       b5 = cast->ft.decl.name.content == "", b6 = cast->ft.decl.has_nested_declarator == false,
-       b7 = cast->kind == cast_kind, b8 = expr->type == expr_type;
-
-  bool res = b2 && b3 && b4 && b5 && b6 && b7 && b8;
-
-  if (!res) {
-    ExpectedMsg = "From IsCast Matcher: ";
-
-    if (!b3) {
-      ExpectedMsg += "ft.flags = 0\n";
-      *result_listener << "got ft.flags = " << to_string(cast->ft.flags) << "\n";
-    }
-    if (!b4) {
-      ExpectedMsg += "ft.decl.components.size() = 0\n";
-      *result_listener << "got ft.decl.components.size() = " << to_string(cast->ft.decl.components.size()) << "\n";
-    }
-    if (!b5) {
-      ExpectedMsg += "ft.decl.name.content = \"\"\n";
-      *result_listener << "got ft.decl.name.content = " << cast->ft.decl.name.content << "\n";
-    }
-    if (!b6) {
-      ExpectedMsg += "ft.decl.has_nested_declarator = false\n";
-      *result_listener << "got ft.decl.has_nested_declarator = " << to_string(cast->ft.decl.has_nested_declarator)
-                       << "\n";
-    }
-    if (!b7) {
-      ExpectedMsg += "cast->kind = " + AST::CastExpression::KindToString(cast_kind) + "\n";
-      *result_listener << "got cast->kind = " << AST::CastExpression::KindToString(cast->kind) << "\n";
-    }
-    if (!b8) {
-      ExpectedMsg += "expr->type = " + AST::NodeToString(expr_type) + "\n";
-      *result_listener << "got expr->type = " << AST::NodeToString(expr->type) << "\n";
-    }
-  }
-
-  return res;
-}
-
-MATCHER_P(IsIdentifier, iden, "") {
-  bool b1 = arg->type == AST::NodeType::IDENTIFIER;
-  if (!b1) {
-    ExpectedMsg += "From IsIdentifier Matcher: NodeType = IDENTIFIER\n";
-    *result_listener << "got NodeType = " << AST::NodeToString(arg->type) << "\n";
-    return 0;
-  }
-
-  bool b2 = arg->template As<AST::IdentifierAccess>()->name.content == iden;
-  if (!b2) {
-    ExpectedMsg += "From IsIdentifier Matcher: ";
-    ExpectedMsg += "Identifier = " + PrintToString(iden) + "\n";
-    *result_listener << "got Identifier = " << arg->template As<AST::IdentifierAccess>()->name.content << "\n";
-    return 0;
-  }
-
-  return 1;
-}
-
-MATCHER_P(IsLiteral, lit, "") {
-  bool b1 = arg->type == AST::NodeType::LITERAL;
-  if (!b1) {
-    ExpectedMsg += "From IsLiteral Matcher: NodeType = LITERAL\n";
-    *result_listener << "got NodeType = " << AST::NodeToString(arg->type) << "\n";
-    return 0;
-  }
-
-  bool b2 = std::get<std::string>(arg->template As<AST::Literal>()->value.value) == lit;
-  if (!b2) {
-    ExpectedMsg += "From IsLiteral Matcher: ";
-    ExpectedMsg += "Literal = " + PrintToString(lit) + "\n";
-    *result_listener << "got Literal = " << std::get<std::string>(arg->template As<AST::Literal>()->value.value)
-                     << "\n";
-    return 0;
-  }
-
-  return 1;
-}
-
-MATCHER_P3(IsBinaryOperation, op, M1, M2, "") {
-  bool b1 = arg->type == AST::NodeType::BINARY_EXPRESSION;
-  if (!b1) {
-    ExpectedMsg += "From IsBinaryOperation Matcher: NodeType = BINARY_EXPRESSION\n";
-    *result_listener << "got NodeType = " << AST::NodeToString(arg->type) << "\n";
-    return 0;
-  }
-
-  auto *binary = arg->template As<AST::BinaryExpression>();
-  bool b2 = binary->operation == op, b3 = ExplainMatchResult(M1, binary->left, result_listener),
-       b4 = ExplainMatchResult(M2, binary->right, result_listener);
-
-  bool res = b2 && b3 && b4;
-  if (!b2) {
-    ExpectedMsg += "From IsBinaryOperation Matcher: ";
-    ExpectedMsg += "Operation = " + ToString(op) + "\n";
-    *result_listener << "got Operation = " << ToString(binary->operation) << "\n";
-  }
-
-  return res;
-}
-
-MATCHER_P2(IsUnaryPostfixOperator, op, M1, "") {
-  bool b1 = arg->type == AST::NodeType::UNARY_POSTFIX_EXPRESSION;
-  if (!b1) {
-    ExpectedMsg = "From IsUnaryPostfixOperator Matcher: NodeType = UNARY_POSTFIX_EXPRESSION\n";
-    *result_listener << "got NodeType = " << AST::NodeToString(arg->type) << "\n";
-    return 0;
-  }
-
-  auto *unary = arg->template As<AST::UnaryPostfixExpression>();
-
-  bool b2 = unary->operation == op, b3 = ExplainMatchResult(M1, unary->operand, result_listener);
-
-  bool res = b2 && b3;
-  if (!b2) {
-    ExpectedMsg = "From IsUnaryPostfixOperator Matcher: ";
-    ExpectedMsg += "Operation = " + ToString(op) + "\n";
-    *result_listener << "got Operation = " << ToString(unary->operation) << "\n";
-  }
-
-  return res;
-}
-
-MATCHER_P2(IsUnaryPrefixOperator, op, M1, "") {
-  bool b1 = arg->type == AST::NodeType::UNARY_PREFIX_EXPRESSION;
-  if (!b1) {
-    ExpectedMsg = "From IsUnaryPrefixOperator Matcher: NodeType = UNARY_PREFIX_EXPRESSION\n";
-    *result_listener << "got NodeType = " << AST::NodeToString(arg->type) << "\n";
-    return 0;
-  }
-
-  auto *unary = arg->template As<AST::UnaryPrefixExpression>();
-  bool b2 = unary->operation == op, b3 = ExplainMatchResult(M1, unary->operand, result_listener);
-
-  bool res = b2 && b3;
-  if (!b2) {
-    ExpectedMsg = "From IsUnaryPrefixOperator Matcher: ";
-    ExpectedMsg += "Operation = " + ToString(op) + "\n";
-    *result_listener << "got Operation = " << ToString(unary->operation) << "\n";
-  }
-
-  return res;
-}
-
-MATCHER_P(IsStatementBlock, stateSize, "") {
-  bool b1 = arg->type == AST::NodeType::BLOCK;
-  if (!b1) {
-    ExpectedMsg = "From IsStatementBlock Matcher: NodeType = BLOCK";
-    *result_listener << "got NodeType = " << AST::NodeToString(arg->type) << "\n";
-    return 0;
-  }
-
-  auto *block = arg->template As<AST::CodeBlock>();
-  bool b2 = block->statements.size() == size_t(stateSize);
-
-  if (!b2) {
-    ExpectedMsg = "From IsStatementBlock Matcher: ";
-    ExpectedMsg = "IsStatementBlock Matcher: Statements Size = " + to_string(stateSize);
-    *result_listener << "Statements Size = " << to_string(block->statements.size()) << "\n";
-  }
-
-  return b2;
-}
-
-MATCHER_P4(IsForLoopWithChildren, M1, M2, M3, M4, ExpectedMsg) {
-  auto *assign = arg->assignment.get();
-  auto *binary = arg->condition.get();
-  auto *unary = arg->increment.get();
-  auto *block = arg->body.get();
-
-  return ExplainMatchResult(M1, assign, result_listener) && ExplainMatchResult(M2, binary, result_listener) &&
-         ExplainMatchResult(M3, unary, result_listener) && ExplainMatchResult(M4, block, result_listener);
 }
 
 TEST(ParserTest, Basics) {
